Add support for reading modem/SIM information from command-line
Since the device name and SIM count are available from the command-line,
add support for loading them from the command-line.  I thought there was
supposed to be a parsed command-line lying around, but the functions to
use are unknown to me.  Suboptimal, but this DOES work.  The fixing
values can still be set in the kernel configuration file.

TODO: The actual value capturing in dirtysanta_fixup_loadcfg() is written
to be moved to kernel initialization time.  This one function call should
be moved there.  In theory there is a race condition, but the likelyhood
of triggering that is extremely low and if triggered they should cause
the same values to be written.

(cherry picked from commit 5a28db6e5c6c9fd453bcd274f5374d0098adbd52)